Refactor Profile and Recipes pages to streamline UI and enhance functionality
- Removed the RecipeRow component from the ProfilePage, simplifying the layout.
- Updated the RecipesPage to include a new RecipeRow component for displaying recent and favorite recipes, as well as recipe collections.
- Improved error handling for user authentication in the RecipesPage.
- Enhanced the overall structure and readability of both pages.
